Reworked the anchors, doc structure and highlights
- added anchor variable to all Front-matter sections
- fixed "Building Your Application" which doesn't highlight at the
moment
- added sub items in the site navigation bar under Welcome
(translations, how to contribute, spread the word!)
- reworked the loop that generates the menu to
    - properly close each <li> tag (I'm surprised that validator.w3.org
doesn't complain about it)
    - create a structure where we have for each section
<li><a>title</a><ul>{<li>subitem</li>…}</ul></li>, which I believe is a
bit cleaner, but I'm not an expert. Also, the final Jekyll output
contains lots of whitespaces, is there a way to automatically clean up
everything and tidy up the compiled document?
- updated the javascript to extract the highlight code to call it not
only when we scroll but when the document is loaded as well, so that
visible items are highlighted from the get go
- added a Credits header since there was an entry in the navigation bar
- changed the css to add the same 10px margin at the bottom of each
section