Remove an extraneous 'to'; rectify an instance of an incorrect verb tense; small adjust to punctuation.
Original text noted that it's "usually a good idea to follow to common patterns". Removed extra 'to'. 

In MVC section the sentence "The model-view-controller (MVC) pattern... let..." should be changed to "The model-view-controller (MVC) pattern...lets..."

Also added a semicolon to a sentence in the Factory section that Strunk and White would agree enhances the clarity of the sentence.