remove singleton from design patterns
While it is a long standing and widely used pattern, I'd be hard pressed to make a case for stateful statics in a modern PHP environment. Even the description includes a solid disclaimer as the final paragraph, which tells me that this pattern is NOT "PHP: The Right Way". Additionally, if this resource is aimed at junior level developers seeking to avoid pitfalls, I feel as though even mentioning these sins from the past, that are usually begrudgingly referred to as "the legacy project" in development organizations, is potentially detrimental.